新版本3.8

+++ BackEnd/src/main/java/com/healthfamily/domain/entity/AiRequestLog.java
@@ -36,10 +36,10 @@ public class AiRequestLog {
     private String modelName;
 
     @Column(name = "input_tokens")
-    private Integer inputTokens;
+    private Long inputTokens;
 
     @Column(name = "output_tokens")
-    private Integer outputTokens;
+    private Long outputTokens;
 
     @Column(name = "latency")
     private Long latency;
